refactor(context): extract shared device detection into a trait
BrowserContext and DeviceContext each carried their own identical copy
of the device-detection plumbing: the nullable service property, the
lazy getDeviceDetectionService() fallback, getRequest() and
getDeviceInfo(). SonarCloud measured the twin block at 36 lines and
failed the quality gate on the PR with
new_duplicated_lines_density 7.3 against a threshold of 3, because the
Rector rewrite in the preceding commit touched lines inside it.

Both copies move into one place, Context\DeviceDetectionAwareTrait,
which both classes now use. Deliberately a single shared definition
rather than a per-class private helper: relocating identical code into
sibling files only moves the duplication and trips the gate again.

The trait lives one namespace above Context\Type on purpose. The PHPat
rule in Tests/Architecture/LayerTest.php asserts that everything in
Context\Type extends AbstractContext, and its namespace selector
matches by prefix, so a trait placed next to the context types would
be pulled into that rule and reported for not extending anything.
Keeping it in Context also leaves both classes extending
AbstractContext directly, exactly as the architecture rule documents,
instead of inserting an intermediate base class purely to share code.

Behaviour is unchanged -- the methods moved verbatim, and no test
subclass overrides any of them (the anonymous test doubles override
getConfValue(), getMatchFromSession() and storeInSession() only).

Verified: 95 unit and 8 functional tests green, PHPStan level 10
without errors including the PHPat architecture rule, CGL clean,
Rector dry-run clean with --clear-cache, php -l on all three files.
The longest common token run between the two classes drops from 116 to
77 tokens, below SonarPHP's 100-token duplication threshold; measured
with a local token-level check that reproduces Sonar's block boundary
on the previous commit exactly.
