+++ Classes/Context/Type/BrowserContext.php
@@ -17,10 +17,9 @@ declare(strict_types=1);
 namespace Netresearch\ContextsDevice\Context\Type;
 
 use Netresearch\Contexts\Context\AbstractContext;
+use Netresearch\ContextsDevice\Context\DeviceDetectionAwareTrait;
 use Netresearch\ContextsDevice\Dto\DeviceInfo;
 use Netresearch\ContextsDevice\Service\DeviceDetectionService;
-use Psr\Http\Message\ServerRequestInterface;
-use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * Context type that matches based on browser name.
@@ -44,7 +43,7 @@ use TYPO3\CMS\Core\Utility\GeneralUtility;
  */
 class BrowserContext extends AbstractContext
 {
-    protected ?DeviceDetectionService $deviceDetectionService = null;
+    use DeviceDetectionAwareTrait;
 
     /**
      * @param array<string, mixed> $arRow Database context row
@@ -91,40 +90,6 @@ class BrowserContext extends AbstractContext
         $bMatch = $this->matchesBrowser($deviceInfo, $configuredBrowsers);
 
         return $this->storeInSession($this->invert($bMatch));
-    }
-
-    /**
-     * Get the device detection service, with lazy initialization fallback.
-     */
-    protected function getDeviceDetectionService(): DeviceDetectionService
-    {
-        if (!$this->deviceDetectionService instanceof DeviceDetectionService) {
-            $this->deviceDetectionService = GeneralUtility::makeInstance(DeviceDetectionService::class);
-        }
-
-        return $this->deviceDetectionService;
-    }
-
-    /**
-     * Get the current HTTP request.
-     */
-    protected function getRequest(): ?ServerRequestInterface
-    {
-        return $GLOBALS['TYPO3_REQUEST'] ?? null;
-    }
-
-    /**
-     * Get device information from the current request.
-     */
-    protected function getDeviceInfo(): ?DeviceInfo
-    {
-        $request = $this->getRequest();
-
-        if (!$request instanceof ServerRequestInterface) {
-            return null;
-        }
-
-        return $this->getDeviceDetectionService()->detectFromRequest($request);
     }
 
     /**
